refactor(shared/system): system 내부 파일 이름 kebab case로 변경
