+++ packages/shared/system/src/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @stylistic/padding-line-between-statements */
-export * from './createComponent';
-export * from './createContext';
+export * from './create-component';
+export * from './create-context';
 export * from './factory';
-export * from './forwardRef';
+export * from './forward-ref';
 export * from './types';
